# web/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swarm Builder",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/graphs/generate": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "Generate Graph Route",
        "description": "Generate, save, and return a graph for ``body.description``.\n\nRaises:\n    HTTPException: 422 when every attempt failed (the detail carries the\n        last review findings as ``problems``); 503 when the resolved\n        route is unmappable, no model is configured, or the generator\n        cannot be imported; 500 when the graph cannot be saved.",
        "operationId": "generate_graph_route_api_graphs_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateGraphRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateGraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "the model could not produce a review-clean graph"
          },
          "502": {
            "description": "the model API refused or failed the request"
          },
          "503": {
            "description": "no usable route or credential, or the generator is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentSpec": {
        "properties": {
          "instructions": {
            "type": "string",
            "title": "Instructions"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools"
          },
          "delegatesTo": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Delegatesto"
          },
          "outputSchema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputschema"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instructions"
        ],
        "title": "AgentSpec",
        "description": "Present on ``kind=\"agent\"`` nodes: instructions, tools, delegation,\nand optional structured output for the generated agent factory.\n\n``output_schema`` is a JSON Schema object. Its values are typed as\n``object`` rather than ``Any``: the *values* are arbitrary, but the\ncontainer is not, and ``Any`` would also silently accept a list or a\nscalar where a schema object was meant."
      },
      "BranchEdge": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "branch",
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "match": {
            "type": "string",
            "title": "Match"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "id",
          "source",
          "target",
          "match"
        ],
        "title": "BranchEdge",
        "description": "One arm of a decision's dispatch, drawn on the canvas from the\n``decision`` node to a branch target. ``match`` must correspond to\none entry in the decision node's :class:`DecisionSpec.branches`."
      },
...
      "DecisionBranch": {
        "properties": {
          "match": {
            "type": "string",
            "title": "Match"
          },
          "targetNodeId": {
            "type": "string",
            "title": "Targetnodeid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match",
          "targetNodeId"
        ],
        "title": "DecisionBranch",
        "description": "One branch of a :class:`DecisionSpec`: a match value and the node\nit dispatches to. ``target_node_id`` must also be the target of a\n``branch`` :class:`SwarmEdge` with a matching ``match`` string."
      },
      "DecisionSpec": {
        "properties": {
          "branches": {
            "items": {
              "$ref": "#/components/schemas/DecisionBranch"
            },
            "type": "array",
            "title": "Branches"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DecisionSpec",
        "description": "Present on ``kind=\"decision\"`` nodes: the branch dispatch table.\n\n**Data does not pass through a decision node.** A branch\ntarget step receives, via ``ctx.inputs``, the *match value returned\nby the decision's source step* -- not the payload that flowed into\nthat source step. Concretely: if the source step returns\n``Literal[\"big\", \"small\"]``, the ``big`` branch target observes\n``ctx.inputs == \"big\"``, never the original pre-classification\nstring. Two consequences that hold regardless of what Phase 1 (Group\n2's review.py) enforces mechanically:\n\n- a branch target's declared :class:`NodeIo.input_type` must equal\n  the **decision's source step's** ``output_type``, never any node\n  further upstream;\n- a workflow that needs the pre-classification payload inside a\n  branch target must carry it through ``State`` explicitly (the\n  classifying step writes a state field the branch target reads);\n  ``ctx.inputs`` will never carry it through the decision.\n\nTwo consequences for the emitted code, both enforced elsewhere: the\nfull ``.branch(...)`` chain must be emitted before the ``add_edge``\nthat targets this decision (``emit_graph.py``), and branch targets'\ndeclared input types must match the source step's output type, not any\nnode further upstream (``review.py``)."
      },
      "DelegateEdge": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "delegate",
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "id",
          "source",
          "target"
        ],
        "title": "DelegateEdge",
        "description": "A delegation link from an orchestrator node to a child agent node.\n\n**Never emitted as an ``add_edge``.** The child is called as an agent\n*tool*, not executed as a graph step; emitting a real edge would run it\ntwice. This edge kind exists purely so the canvas can draw the\nrelationship and the Inspector can show it: ``emit_graph.py`` skips\nevery ``delegate`` edge, and Phase 1 rejects a node that is both a\n``delegate`` target and a ``seq``/``branch`` target."
      },
...
      "FanoutEdge": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "fanout",
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "joinNodeId": {
            "type": "string",
            "title": "Joinnodeid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "id",
          "source",
          "target",
          "joinNodeId"
        ],
        "title": "FanoutEdge",
        "description": "One arm of a multi-successor fan-out.\n\n``join_node_id`` names the ``join`` node this arm must converge on.\nIt is required rather than optional because a plain multi-successor\n``add_edge`` pair is silently lossy: ``build()`` accepts it, and only\none arbitrary branch's value survives without an explicit join."
      },
      "FindingOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "nodeIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nodeids"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "nodeIds"
        ],
        "title": "FindingOut",
        "description": "One review finding, as the canvas Inspector displays it."
      },
      "GenerateGraphRequest": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "graphId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graphid",
            "description": "Reuse an existing graph id (replace its document); default mints a new one."
          },
          "modelOverride": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelSelection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "GenerateGraphRequest",
        "description": "Body of ``POST /api/graphs/generate``."
      },
      "GenerateGraphResponse": {
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/SwarmGraph"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "type": "array",
            "title": "Warnings"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "model": {
            "$ref": "#/components/schemas/GenerateModelOut"
          },
          "dryRun": {
            "type": "boolean",
            "title": "Dryrun"
          }
        },
        "type": "object",
        "required": [
          "graph",
          "warnings",
          "attempts",
          "model",
          "dryRun"
        ],
        "title": "GenerateGraphResponse"
      },
      "GenerateModelOut": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "source"
        ],
        "title": "GenerateModelOut"
      },
...
      "JoinEdge": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "join",
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "id",
          "source",
          "target"
        ],
        "title": "JoinEdge",
        "description": "An edge from a fan-out arm's target into the ``join`` node."
      },
      "JoinSpec": {
        "properties": {
          "reducer": {
            "type": "string",
            "enum": [
              "list_append",
              "list_extend",
              "dict_update",
              "sum"
            ],
            "title": "Reducer"
          },
          "initialFactory": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "list",
                  "dict",
                  "int"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Initialfactory"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reducer"
        ],
        "title": "JoinSpec",
        "description": "Present on ``kind=\"join\"`` nodes: the fan-in reducer.\n\n``initial_factory`` names the Python builtin the emitter passes as\n``builder.join(..., initial_factory=...)``. When unset, the emitter\nderives it from ``reducer`` (``list_append``/``list_extend`` ->\n``list``, ``dict_update`` -> ``dict``, ``sum`` -> ``int``); set it\nexplicitly only to override that default."
      },
...
      "ModelSelection": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "reasoningEffort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoningeffort"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "title": "ModelSelection",
        "description": "One inherited route, resolved to something PydanticAI can\nconstruct (facts 21-24). Absent on :class:`SwarmGraph` means\n\"inherit the resolved default\" rather than \"no model\"."
      },
...
      "NodeIo": {
        "properties": {
          "inputType": {
            "type": "string",
            "enum": [
              "str",
              "json",
              "list[str]"
            ],
            "title": "Inputtype"
          },
          "outputType": {
            "type": "string",
            "enum": [
              "str",
              "json",
              "list[str]"
            ],
            "title": "Outputtype"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "inputType",
          "outputType"
        ],
        "title": "NodeIo",
        "description": "A node's declared input/output port types.\n\nPhase 1 review uses these to reject a port-type mismatch across an\nedge and, for a decision's branch targets, to enforce that the\ntarget's ``input_type`` equals the decision's *source step's*\n``output_type`` -- see :class:`DecisionSpec` for why the source step is\nthe relevant node rather than the decision itself."
      },
...
      "Position": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "Position",
        "description": "Canvas coordinates for a node. Presentation-only: never read by\nthe emitter, kept purely so the canvas restores node placement."
      },
      "ProgrammaticSpec": {
        "properties": {
          "needs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Needs"
          },
          "signatureHint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signaturehint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProgrammaticSpec",
        "description": "Present on ``kind=\"programmatic\"`` nodes: plain-Python steps with\nno agent. ``needs`` are extra package requirements the scaffolder\nunions into the generated ``pyproject.toml``; ``signature_hint`` is\nfree-text guidance the fill agent reads (never executed or parsed)."
      },
...
      "SeqEdge": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "seq",
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "id",
          "source",
          "target"
        ],
        "title": "SeqEdge",
        "description": "A plain sequential edge: ``builder.add_edge(source, target)``."
      },
...
      "StateField": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "enum": [
              "str",
              "json",
              "list[str]"
            ],
            "title": "Type"
          },
          "default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "StateField",
        "description": "One field of the generated workflow's ``@dataclass State``.\n\n``type`` goes through :data:`PORT_TYPE_ANNOTATIONS` like any other\nport type when the emitter writes the dataclass field.\n``default`` is literal Python source for the field's default value\n(e.g. ``'\"\"'`` or ``\"0\"``); ``None`` means the emitter writes a bare\n``name: type`` line with no default."
      },
      "SwarmGraph": {
        "properties": {
          "version": {
            "type": "integer",
            "const": 1,
            "title": "Version",
            "default": 1
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "entryNodeId": {
            "type": "string",
            "title": "Entrynodeid"
          },
          "exitNodeId": {
            "type": "string",
            "title": "Exitnodeid"
          },
          "stateFields": {
            "items": {
              "$ref": "#/components/schemas/StateField"
            },
            "type": "array",
            "title": "Statefields"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SwarmNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SeqEdge"
                },
                {
                  "$ref": "#/components/schemas/BranchEdge"
                },
                {
                  "$ref": "#/components/schemas/FanoutEdge"
                },
                {
                  "$ref": "#/components/schemas/JoinEdge"
                },
                {
                  "$ref": "#/components/schemas/DelegateEdge"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "branch": "#/components/schemas/BranchEdge",
                  "delegate": "#/components/schemas/DelegateEdge",
                  "fanout": "#/components/schemas/FanoutEdge",
                  "join": "#/components/schemas/JoinEdge",
                  "seq": "#/components/schemas/SeqEdge"
                }
              }
            },
            "type": "array",
            "title": "Edges"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "entryNodeId",
          "exitNodeId",
          "updatedAt"
        ],
        "title": "SwarmGraph",
        "description": "The complete canvas save file: nodes, edges, state shape, and the\noptional per-workflow model override. ``version`` is fixed at ``1``\nfor this schema generation; a future incompatible shape change ships\nas ``version=2`` plus a migration, not a silent field rename."
      },
      "SwarmNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "agent",
              "programmatic",
              "decision",
              "join"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          },
          "template": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "chat",
                  "orchestrator",
                  "websearch"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Template"
          },
          "io": {
            "$ref": "#/components/schemas/NodeIo"
          },
          "reads": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reads"
          },
          "writes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Writes"
          },
          "agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "programmatic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgrammaticSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "join": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JoinSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "intent",
          "position",
          "io"
        ],
        "title": "SwarmNode",
        "description": "One canvas node.\n\n``id`` is emitted verbatim as the ``node_id=`` argument to\n``builder.step``/``builder.decision``/``builder.join`` and as the step\nmodule's filename, so it must already be a valid, unique Python\nidentifier. The frontend is responsible for deriving it from ``title``\nvia :func:`swarm_builder.slugify.slugify_titles` and keeping it stable\nacross edits: duplicate titles deduplicate with a numeric suffix, and\nthe title-to-id mapping must stay stable across recompiles or a\nrecompile would rename step modules instead of updating them.\n\nExactly one of ``agent`` / ``programmatic`` / ``decision`` / ``join``\nis expected to be set, matching ``kind`` -- enforced by Phase 1\n(``review.py``), not here, since that cross-field rule belongs with the\nrest of the structural graph validation."
      },
...
    }
  }
}